Move QueryParameters to dynatrace package
